ai-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0
